backend/services/word_parser/requirements.txt
# Core web framework
fastapi
uvicorn[standard]

# For parsing Word documents (.docx)
python-docx
